requirements.txt
# Core Dependencies
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Vector Store
chromadb>=0.4.0
tiktoken>=0.5.2

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0
kaleido>=0.2.1

# Web Interface
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: For enhanced features
scikit-learn>=1.3.0
